% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_cor.R
\name{matrix_cor}
\alias{matrix_cor}
\title{matrix_cor}
\usage{
matrix_cor(X, Y, method = "pearson")
}
\arguments{
\item{X}{A matrix}

\item{Y}{A matrix}

\item{method}{A string}
}
\value{
For `method = "pearson"`, returns an object the same type of `X`
containing the correlation between all the columns of `X` and all the
columns of `Y`.
}
\description{
compute the correlation between the columns of X and the columns $Y$.
}
\examples{
matrix_cor(rnorm(20), rnorm(20), method="pearson")
matrix_cor(matrix (c(1, 2, 2, 1), 2, 2), c(1,2), method="kendall")
}
