% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_cor.R
\name{matrix_cor}
\alias{matrix_cor}
\title{Correlation matrices}
\usage{
matrix_cor(X, Y = NULL, method = "pearson")
}
\arguments{
\item{X}{A matrix or a vector}

\item{Y}{A matrix or a vector}

\item{method}{A string}
}
\value{
For \code{method = "pearson"}, returns a matrix containing the pearson
correlation coefficient between all the columns of \code{X} and all the columns
of \code{Y}.

For \code{method = "kendall"}, returns a \emph{vector} containing the kendall rank
correlation between columns of \code{X} and the corresponding columns of \code{Y}.
}
\description{
\code{matrix_cor()} computes the correlation between the columns of
\code{X} and the columns of \code{Y}.
}
\details{
To learn more about this package, start with the vignettes:
\code{browseVignettes(package = "vectorizedCorrelations")}
}
\examples{
matrix_cor(rnorm(20), rnorm(20), method="pearson")
matrix_cor(matrix (c(1, 2, 2, 1), 2, 2), c(1,2), method="kendall")
}
